Cast to narrow type from Vec_Get_Size.
Cast from size_t to uint32_t to avoid -Wconversion warnings.  In each of
these cases, the number of elements in the Vector should be small.

+++ core/Lucy/Search/PhraseMatcher.c
@@ -46,7 +46,7 @@ PhraseMatcher_init(PhraseMatcher *self, Similarity *similarity, Vector *plists,
     ivars->more             = true;
 
     // Extract PostingLists out of Vector into local C array for quick access.
-    ivars->num_elements = Vec_Get_Size(plists);
+    ivars->num_elements = (uint32_t)Vec_Get_Size(plists);
     ivars->plists = (PostingList**)MALLOCATE(
                        ivars->num_elements * sizeof(PostingList*));
     for (size_t i = 0; i < ivars->num_elements; i++) {
